assets/2025/target-checker/requirements.txt
requests
wolfsoftware.pushover